archival: Use read-write fence in the ntp_archiver
Previously, the `read_write_fence` command was added to the
`archival_metadata_stm` to ensure consistency. The command operates as
follows:
- The STM stores the offset of the last applied command.
- The `read_write_fence` command is initialized with the last applied
  offset and is included as the first entry in the archival STM record
  batch.
- When the batch is replicated and eventually applied to the STM's
  in-memory state, the `read_write_fence` compares its stored offset
  with the STM's last applied offset.
  - If the offsets do not match, all remaining commands in the batch
    are discarded.
  - If the offsets match, the remaining commands are applied.

This mechanism ensures that if any changes to the STM's in-memory state
occur after the ntp_archiver reads the last applied offset, the batch is
aborted. This requires the `read_write_fence` to always be the first
command in the record batch.

Currently, the `ntp_archiver` does not emit the `read_write_fence`
command. This commit addresses that by implementing the following logic:
1. The `ntp_archiver` reads the last applied offset from the STM at the
   start of a new upload iteration.
2. It searches for an upload candidate and performs the upload.
3. The `ntp_archiver` replicates metadata describing the upload,
   including a `read_write_fence` command in the record batch.
4. The `read_write_fence` is initialized using the last applied offset
   cached at the beginning.

This ensures that the configuration batch generated by the `ntp_archiver`
can only be applied to the STM if its state remains unchanged. Any
concurrent changes to the STM state will prevent the batch from being
applied.

This utility ensures that if any changes to the STM are made while
segments are uploaded to S3, there will be no metadata conflict. The
same holds true if a portion of the log is replayed by the STM twice.
Every configuration batch can only be applied to a specific version of
the archival STM state. This makes the archival STM effectively
idempotent.

Note, that there is no way for the 'ntp_archiver' to replay any log
messages twice or make any concurrent modifications to the archival STM.
This is only possible if we have some bug. We're not relying on this
mechanism for correctness and only using it as a safety net.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>