cluster: Fix archival metadata STM snapshot
The snapshot doesn't serialize the applied-offset field of the manifest.
This commit fixes this.

The field is used by the fencing mechanism. Prior to this commit (and
the PR that adds this commit) the field was write only. It was updated
and never actually used because we don't have any code in the write path
that generates read_write_fence commands for the archival STM.

Before the fencing mechanism could be used the snapshot has to be fixed,
otherwise the replicas may diverge.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>