cluster: Remove consistency check from the archival STM
The check is no longer needed because we're using fencing mechanism to
prevent concurrency. The check is actually performed when the segment is
uploaded. The removed check was used as an optimistic concurrency
control mechanism and is now replaced by the RW-fence.

Also, deprecate the error code that STM uses to communicate the
consistency violation to the caller. The error code is not used anywhere
else.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>