archival: Refactor wait_uploads method
Split method into two parts:
- wait_uploads_complete
- replicate_archival_metadata

The first method is waiting until all uploads are finished. But unlike
wait_uploads it doesn't replicate the metadata. Instead of doing this it
just returns all information needed to replicate it.

The second method takes the output of the wait_uploads_complete and
replicates it.

Finally, the wait_uploads method is now implemented using this two new
methods.

This is a preparation for the next commit. Currently, when we're
uploading both compacted and non-compacted segments in the same
iteration of the upload loop they both replicated independently. This
worked fine before (with the exception of the race condition on
_projected_manifest_clean_at). But with the fencing mechanism one of
this wait_uploads calls will fail to replicate because of the fence.
Currently, the same fence is used to invoke wait_uploads for compacted
and non-compacted segments. The solution to this problem is to use
wait_uploads_complete to wait until both compacted and non-compacted
uploads are completed and then use a single replicate_archival_metadata
call. This call will be able to use the fence value correctly.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>